requirements.txt
python-telegram-bot = "13.2"
python-dotenv = "^1.0.0"
requests = "2.28.1"
